[v0.88][tools] Split pr_cmd_inline finish tests by concern (Closes #1717)
