+++ adl/src/cli/tests/pr_cmd_inline/finish/mod.rs
@@ -1,0 +1,5 @@
+use super::*;
+
+mod arg_render;
+mod guardrails;
+mod publication;
